Let `Parser` handle multiple pages correctly.
In the process, make it so that `Fetcher` handles non-absolute URLs better. It
can now deal with hash URLs as well and will resolve absolute URLs and hash URLs
to the URL of the page that URL was found on, instead of the 'start' URL, which
may be on a different domain.

Wrote a test to verify that the new changes work correctly in ParserTest`.