Add some convenience methods to `QueryTranslator`.
This makes it easier to translate a query in general and adds support for adding
schemes to queries to make absolutely sure that a query is of a certain type.
That is 'css:a.class' is now supported, as is 'xpath://a[@id="id"]' and those
two without the schemes are too, but a heuristic is used to determine if a query
is CSS or XPath in that case.