Use `QueryTranslator` so that config can have CSS queries.
The `Parser` now uses the `QueryTranslator`, enabling CSS queries that are
specified in the configuration to find interesting elements. They can also be
used for 'get' and 'next', but those require text values to be returned, while
CSS can only select DOM nodes.

I also changed one query in the test configuration to be CSS. All tests still
pass, so it appears to work correctly.