Use Composer for dependencies, fix tests
This gets rid of the Git submodule and instead uses Composer, which is
easier to maintain. Because it has been so long, the tests broke. They
are fixed by this commit. They are also more robust, in the sense that
not having the cURL extension will fail with an informative error
messages, instead of crashing all tests. Finally, somehow I had to apply
fixes to the code. Not sure why this is the case.